+++ .jupyter/lab/workspaces/lab-a511.jupyterlab-workspace
@@ -1,1 +1,1 @@
-{"data":{"@jupyterlab/settingeditor-extension:plugin":{"sizes":[0.12448132780082988,0.8755186721991701],"container":{"plugin":"@jupyterlab/notebook-extension:tracker","sizes":[0.4774703557312253,0.5225296442687747]}},"layout-restorer:data":{"main":{"dock":{"type":"tab-area","currentIndex":0,"widgets":["notebook:Untitled.ipynb"]},"mode":"multiple-document","current":"notebook:Untitled.ipynb"},"left":{"collapsed":false,"current":"filebrowser","widgets":["filebrowser","running-sessions","command-palette","jp-property-inspector","tab-manager","juputerlab-toc","extensionmanager.main-view"]},"right":{"collapsed":true,"widgets":[]}},"file-browser-github-filebrowser:cwd":{"path":"GitHub:"},"variableinspector:variableinspector":{"data":null},"notebook:Untitled.ipynb":{"data":{"path":"Untitled.ipynb","factory":"Notebook"}}},"metadata":{"id":"/lab"}}
+{"data":{"@jupyterlab/settingeditor-extension:plugin":{"sizes":[0.12448132780082988,0.8755186721991701],"container":{"plugin":"@jupyterlab/notebook-extension:tracker","sizes":[0.4774703557312253,0.5225296442687747]}},"layout-restorer:data":{"main":{"dock":{"type":"tab-area","currentIndex":0,"widgets":[]},"mode":"multiple-document"},"left":{"collapsed":false,"current":"extensionmanager.main-view","widgets":["filebrowser","running-sessions","command-palette","jp-property-inspector","tab-manager","juputerlab-toc","extensionmanager.main-view"]},"right":{"collapsed":true,"widgets":[]}},"file-browser-github-filebrowser:cwd":{"path":"GitHub:"},"variableinspector:variableinspector":{"data":null}},"metadata":{"id":"/lab"}}
